consolidate app files & update sources to ESM
