+++ components/mongodb/mongodb.app.mjs
@@ -1,4 +1,5 @@
 import mongoose from "mongoose";
+import { MongoClient } from "mongodb";
 
 export default {
   type: "app",
@@ -17,8 +18,95 @@ export default {
       type: "string",
       description: "The document ID",
     },
+    database: {
+      type: "string",
+      label: "Database",
+      description: "A MongoDB database",
+      async options() {
+        const client = await this.getClient();
+        const databases = await this.listDatabases(client);
+        await client.close();
+        await new Promise((resolve) => setTimeout(resolve, 3000));
+        return databases.map((db) => {
+          return db.name;
+        });
+      },
+    },
+    sourceCollection: {
+      type: "string",
+      label: "Collection",
+      description: "A MongoDB collection",
+      async options(opts) {
+        const client = await this.getClient();
+        const db = this.getDatabase(client, opts.database);
+        const collections = await this.listCollections(db);
+        await client.close();
+        await new Promise((resolve) => setTimeout(resolve, 3000));
+        return collections.map((collection) => {
+          return collection.name;
+        });
+      },
+    },
+    sourceDocument: {
+      type: "string",
+      label: "Document",
+      description: "Select a document from the MongoDB database",
+      async options(opts) {
+        const client = await this.getClient();
+        const db = this.getDatabase(client, opts.database);
+        const collection = this.getCollection(db, opts.collection);
+        const documents = await this.listDocuments(collection);
+        await client.close();
+        await new Promise((resolve) => setTimeout(resolve, 3000));
+        return documents.map((doc) => {
+          return doc._id;
+        });
+      },
+    },
   },
   methods: {
+    _getConnectionUrl(replSet = true) {
+      const {
+        username,
+        password,
+        hostname,
+      } = this.$auth;
+      const protocol = replSet
+        ? "mongodb+srv"
+        : "mongodb";
+      return `${protocol}://${username}:${password}@${hostname}`;
+    },
+    async getClient() {
+      const url = this._getConnectionUrl();
+      try {
+        return await MongoClient.connect(url);
+      } catch (err) {
+        if (err.code === "ENOTFOUND") {
+          // Last attempt to connect to the server
+          const url = this._getConnectionUrl(false);
+          return MongoClient.connect(url);
+        }
+        throw err;
+      }
+    },
+    getDatabase(client, dbName) {
+      return client.db(dbName);
+    },
+    getCollection(db, collectionName) {
+      return db.collection(collectionName);
+    },
+    async listDatabases(client) {
+      const db = this.getDatabase(client, "admin");
+      const adminDb = await db.admin();
+      const { databases } = await adminDb.listDatabases();
+      return databases;
+    },
+    async listCollections(db) {
+      return await db.listCollections().toArray();
+    },
+    async listDocuments(collection) {
+      return await collection.find().toArray();
+    },
     async connect() {
       const {
         username,
