+++ components/mongodb/sources/new-database/new-database.mjs
@@ -1,6 +1,6 @@
-const common = require("../common.js");
+import common from "../common.mjs";
 
-module.exports = {
+export default {
   ...common,
   key: "mongodb-new-database",
   name: "New Database",
